backend/requirements.txt
# asgiref==3.8.1
# certifi==2024.8.30
# cffi==1.17.1
# channels==4.2.0
# charset-normalizer==3.4.0
# class-registry==2.1.2
# cryptography==44.0.0
# Django==5.1.4
# django-comment-migrate==0.1.7
# django-cors-headers==4.6.0
# django-filter==24.3
# django-ranged-response==0.2.0
# django-restql==0.15.4
# django-simple-captcha==0.6.0
# djangorestframework==3.15.2
# djangorestframework-simplejwt==5.3.1
# drf-yasg==1.21.8
# et_xmlfile==2.0.0
# filters==1.3.2
# idna==3.10
# inflection==0.5.1
# openpyxl==3.1.5
# packaging==24.2
# pillow==11.0.0
# pycparser==2.22
# PyJWT==2.10.1
# PyMySQL==1.1.1
# pyPEG2==2.15.2
# python-dateutil==2.9.0.post0
# pytz==2024.2
# PyYAML==6.0.2
# regex==2024.11.6
# requests==2.32.3
# six==1.17.0
# sqlparse==0.5.3
# ua-parser==1.0.0
# ua-parser-builtins==0.18.0.post1
# uritemplate==4.1.1
# urllib3==2.2.3
# user-agents==2.2.0

asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
class-registry==2.1.2
cryptography==44.0.0
Django==5.1.4
django-comment-migrate==0.1.7
django-cors-headers==4.6.0
django-filter==24.3
django-ranged-response==0.2.0
django-restql==0.15.4
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
filters==1.3.2
idna==3.10
inflection==0.5.1
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyPEG2==2.15.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0